C and other Super I/O cosmetic fixes.
 - Random coding style, whitespace and cosmetic fixes.
 
 - Consistently use the same spacing and 4-hexdigit port number format
   in the pnp_dev_info[] arrays.
 
 - Drop dead/unused code and less useful comments.
 
 - Add missing "(C)" characters and copyright years.
 
 - Shorten and simplify some code snippets.
 
 - Use u8/u16/etc. everywhere.

Acked-by: Uwe Hermann <uwe@hermann-uwe.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@6073 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/superio/nsc/pc8374/pc8374_early_init.c
@@ -22,30 +22,31 @@
 #include <arch/romcc_io.h>
 #include "pc8374.h"
 
-/* things that Must Be Done to get this chip working */
-/* Straight from the data book */
-static void pc8374_enable(unsigned iobase, u8 *init)
+static void pc8374_enable(u16 iobase, u8 *init)
 {
 	u8 val, count;
+
 	outb(0x29, iobase);
-	val = inb(iobase+1);
+	val = inb(iobase + 1);
 	val |= 0x91;
-	outb(val, iobase+1);
-	for(count = 0; count < 255; count++)
-		if (inb(iobase+1) == 0x91)
+	outb(val, iobase + 1);
+
+	for (count = 0; count < 255; count++)
+		if (inb(iobase + 1) == 0x91)
 			break;
-	for(;*init; init++) {
+
+	for (; *init; init++) {
 		outb(*init, iobase);
-		val = inb(iobase+1);
+		val = inb(iobase + 1);
 		init++;
 		val &= *init;
 		init++;
 		val |= *init;
-		outb(val, iobase+1);
+		outb(val, iobase + 1);
 	}
 }
 
-static void pc8374_enable_dev(device_t dev, unsigned iobase)
+static void pc8374_enable_dev(device_t dev, u16 iobase)
 {
 	pnp_set_logical_device(dev);
 	pnp_set_enable(dev, 0);
