+++ src/superio/nsc/pc87309/pc87309_early_serial.c
@@ -21,7 +21,7 @@
 #include <arch/romcc_io.h>
 #include "pc87309.h"
 
-static void pc87309_enable_serial(device_t dev, unsigned int iobase)
+static void pc87309_enable_serial(device_t dev, u16 iobase)
 {
 	pnp_set_logical_device(dev);
 	pnp_set_enable(dev, 0);
